+++ ginc/src/ast/expr/range.rs
@@ -15,7 +15,7 @@ impl Range {
     }
 }
 
-pub fn int_range<'t, I>() -> impl Parser<'t, I, std::ops::Range<i64>, ParserError<'t>> + Clone
+pub fn int_range<'t, I>() -> impl Parser<'t, I, std::ops::Range<i128>, ParserError<'t>> + Clone
 where
     I: ValueInput<'t, Token = Token<'t>, Span = SimpleSpan>,
 {
